move 'symmetricDifferenceWith' tests to separate file
