+++ types/ramda/ramda-tests.ts
@@ -423,26 +423,6 @@ R.times(i, 5);
     const a: number[] = R.without([1, 2], [1, 2, 1, 3, 4]); // => [3, 4]
 };
 
-/*****************************************************************
- * Relation category
- */
-() => {
-    const eqA = R.eqBy(R.prop("a"));
-    const l1  = [{a: 1}, {a: 2}, {a: 3}, {a: 4}];
-    const l2  = [{a: 3}, {a: 4}, {a: 5}, {a: 6}];
-    R.symmetricDifferenceWith(eqA, l1, l2); // => [{a: 1}, {a: 2}, {a: 5}, {a: 6}]
-    R.symmetricDifferenceWith(eqA)(l1, l2); // => [{a: 1}, {a: 2}, {a: 5}, {a: 6}]
-    // const c: (a: any[]) => any[] = R.symmetricDifferenceWith(eqA)(l1); // => [{a: 1}, {a: 2}, {a: 5}, {a: 6}]
-};
-
-() => {
-    const eqL = R.eqBy<string, number>(s => s.length);
-    const l1 = ['bb', 'ccc', 'dddd'];
-    const l2 = ['aaa', 'bb', 'c'];
-    R.symmetricDifferenceWith(eqL, l1, l2); // => ['dddd', 'c']
-    R.symmetricDifferenceWith(eqL)(l1, l2); // => ['dddd', 'c']
-};
-
 () => {
     interface A {
         a: number;
