+++ types/ramda/tsconfig.json
@@ -208,6 +208,7 @@
         "test/startsWith-tests.ts",
         "test/subtract-tests.ts",
         "test/sum-tests.ts",
-        "test/symmetricDifference-tests.ts"
+        "test/symmetricDifference-tests.ts",
+        "test/symmetricDifferenceWith-tests.ts"
     ]
 }
